# atoma-proxy/docs/openapi.yml
openapi: 3.1.0
info:
  title: atoma-proxy
  description: ''
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
servers:
- url: http://localhost:8080
paths:
...
  /v1/embeddings:
    post:
      tags:
      - Embeddings
      summary: Create embeddings
      description: |-
        This endpoint follows the OpenAI API format for generating vector embeddings from input text.
        The handler receives pre-processed metadata from middleware and forwards the request to
        the selected node.

        Note: Authentication, node selection, and initial request validation are handled by middleware
        before this handler is called.

        # Arguments
        * `metadata` - Pre-processed request metadata containing node information and compute units
        * `state` - The shared proxy state containing configuration and runtime information
        * `headers` - HTTP headers from the incoming request
        * `payload` - The JSON request body containing the model and input text

        # Returns
        * `Ok(Response)` - The embeddings response from the processing node
        * `Err(AtomaProxyError)` - An error status code if any step fails

        # Errors
        * `INTERNAL_SERVER_ERROR` - Processing or node communication failures
      operationId: embeddings_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmbeddingRequest'
        required: true
      responses:
        '200':
          description: Embeddings generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEmbeddingResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
  /v1/confidential/embeddings:
    post:
      tags:
      - Confidential Embeddings
      summary: Create confidential embeddings
      description: |-
        This endpoint follows the OpenAI API format for generating vector embeddings from input text,
        but with confidential processing (through AEAD encryption and TEE hardware).
        The handler receives pre-processed metadata from middleware and forwards the request to
        the selected node.

        Note: Authentication, node selection, initial request validation and encryption
        are handled by middleware before this handler is called.

        # Arguments
        * `metadata` - Pre-processed request metadata containing node information and compute units
        * `state` - The shared proxy state containing configuration and runtime information
        * `headers` - HTTP headers from the incoming request
        * `payload` - The JSON request body containing the model and input text

        # Returns
        * `Ok(Response)` - The embeddings response from the processing node
        * `Err(AtomaProxyError)` - An error status code if any step fails

        # Errors
        * `INTERNAL_SERVER_ERROR` - Processing or node communication failures
      operationId: confidential_embeddings_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmbeddingRequest'
        required: true
      responses:
        '200':
          description: Confidential embeddings generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEmbeddingResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    CreateEmbeddingRequest:
      allOf:
      - $ref: '#/components/schemas/EmbeddingInput'
        description: |-
          Input text to get embeddings for. Can be a string or array of strings.
          Each input must not exceed the max input tokens for the model
      - type: object
        required:
        - model
        properties:
          dimensions:
            type:
            - integer
            - 'null'
            format: int32
            description: |-
              The number of dimensions the resulting output embeddings should have.
              Only supported in text-embedding-3 models.
            minimum: 0
          encoding_format:
            type:
            - string
            - 'null'
            description: |-
              The format to return the embeddings in. Can be "float" or "base64".
              Defaults to "float"
          model:
            type: string
            description: ID of the model to use.
          user:
            type:
            - string
            - 'null'
            description: A unique identifier representing your end-user, which can help OpenAI to monitor and detect abuse.
      description: Request object for creating embeddings
    CreateEmbeddingResponse:
      type: object
      description: Response object from creating embeddings
      required:
      - object
      - model
      - data
      - usage
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/EmbeddingObject'
          description: List of embedding objects
        model:
          type: string
          description: The model used for generating embeddings
        object:
          type: string
          description: The object type, which is always "list"
        usage:
          $ref: '#/components/schemas/EmbeddingUsage'
          description: Usage statistics for the request
...
    EmbeddingInput:
      oneOf:
      - type: string
      - type: array
        items:
          type: string
      description: Input types for embeddings request
    EmbeddingObject:
      type: object
      description: Individual embedding object in the response
      required:
      - object
      - embedding
      - index
      properties:
        embedding:
          type: array
          items:
            type: number
            format: float
          description: The embedding vector
        index:
          type: integer
          description: Index of the embedding in the list of embeddings
          minimum: 0
        object:
          type: string
          description: The object type, which is always "embedding"
    EmbeddingUsage:
      type: object
      description: Usage information for the embeddings request
      required:
      - prompt_tokens
      - total_tokens
      properties:
        prompt_tokens:
          type: integer
          format: int32
          description: Number of tokens in the prompt
          minimum: 0
        total_tokens:
          type: integer
          format: int32
          description: Total tokens used in the request
          minimum: 0
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...